+++ src/Routes/Routes.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Home from "../Pages/Home";
-import Nav from "../Pages/Nav";
+import Nav from "../Pages/Navv";
 import Login from "../Pages/Login";
 import Signup from "../Pages/Signup";
 import Measurments from "../Pages/Measurments";
